Add files via upload
Here I have done analysis of loan providing companies, where they find it hard to give loans to the people due to their insufficient or non-existent credit history. Because of that, some consumers use it as their advantage by becoming a defaulter. I have tried to figure out various types of loans to urban customers with the help of finance department. I have use EDA to analyse the patterns present in the data to ensure that the applicants capable of repaying the loan are not rejected.